requirements.txt
redisvl
numpy<2.0
sentence-transformers
python-dotenv
pydantic-ai-slim[openai]